added my file

+++ calculator/calculator.js
@@ -8,14 +8,14 @@ function calculator(num){
 	}
 	else {
 		if(num.value == "+" || num.value == "-" || num.value == "*" || num.value == "/" || num.value == "%" ){
-			first_number = parseInt(document.getElementById("display").value);
+			first_number = parseFloat(document.getElementById("display").value);
 			operator = num.value;
 			document.getElementById("display").value = "";
 		}
 		
 		else if(num.value == "="){
 			
-			second_number = parseInt(document.getElementById("display").value);
+			second_number = parseFloat(document.getElementById("display").value);
 			if(operator == "+"){
 				
 				document.getElementById("display").value = first_number + second_number;
@@ -33,6 +33,7 @@ function calculator(num){
 			else if(operator == "%"){
 				document.getElementById("display").value = first_number % second_number;
 			}
+			
 				
 			
 		}
